+++ init.rb
@@ -1,5 +1,5 @@
 require_plugin 'tog_core'
-require_plugin 'acts_as_state_machine'
+#require_plugin 'acts_as_state_machine'
 require_plugin 'seo_urls'
 require_plugin 'paperclip'
 require_plugin 'acts_as_shareable'