+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Edge
 ----
 
 * Replaced own sharing code by acts_as_shareable plugin. Please, install acts_as_shareable and it's migration before running plugins migrations.
+* Acts_as_state_machine code replaced by AASM gem
 
 0.5.4
 ----